requirements.txt
streamlit==1.28.1
openai==1.3.7
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
pytest==7.4.3 